+++ src/Zafiro/Reactive/ObservableMixin.cs
@@ -10,9 +10,12 @@ using System.Reactive.Linq;
 using DynamicData;
 using DynamicData.Binding;
 using ReactiveUI;
+using Zafiro;
 using Zafiro.Actions;
+using Zafiro.Mixins;
+using Zafiro.Reactive;
 
-namespace Zafiro.Mixins;
+namespace Zafiro.Reactive;
 
 public static class ObservableMixin
 {
@@ -168,7 +171,7 @@ public static class ObservableMixin
     /// <param name="collection"></param>
     /// <returns></returns>
     public static IDisposable UpdateCollectionWhenSomeOtherCollectionObservableChanges<T, TItem>(
-        this T parent, 
+        this T parent,
         Expression<Func<T, ReadOnlyObservableCollection<TItem>>> selector, out ReadOnlyObservableCollection<TItem> collection) where TItem : notnull where T : ReactiveObject
     {
         CompositeDisposable disposable = new();
@@ -218,21 +221,5 @@ public static class ObservableMixin
             .DisposeWith(disposable);
 
         return disposable;
-    }
-
-    public static IObservable<byte> ProgressDo(this IObservable<byte> bytes, long length, Action<LongProgress> action, IScheduler? scheduler = null)
-    {
-        long totalBytes = 0;
-
-        return bytes
-            .Buffer(TimeSpan.FromSeconds(1), scheduler: scheduler ?? Scheduler.Default)
-            .Where(list => list.Any())
-            .Do(list =>
-            {
-                totalBytes += list.Count; 
-                var progress = new LongProgress(totalBytes, length);
-                action(progress);
-            })
-            .SelectMany(listOfBytes => listOfBytes);
     }
 }
